refactor(Problem4): 클래스 분리
