+++ src/main/java/onboarding/problem4/Constant.java
@@ -1,0 +1,7 @@
+package onboarding.problem4;
+
+public class Constant {
+    static final String EXCEPTION = "[ERROR]";
+    static final int LENGTH_MAX = 1000;
+    static final int LENGTH_MIN = 1;
+}
