+++ src/main/java/onboarding/problem4/Controller.java
@@ -1,0 +1,11 @@
+package onboarding.problem4;
+
+public class Controller {
+    public static String getSolution(String word) {
+        if (Validation.check(word)) {
+            String result = Word.changeWord(word);
+            return result;
+        }
+        return Constant.EXCEPTION;
+    }
+}
